rename tasks.js to not be confused with tasks.js
