+++ lib/plugins/digging.js
@@ -1,5 +1,5 @@
 const { performance } = require('perf_hooks')
-const { taskCreateDone, tasks } = require('../tasks')
+const { taskCreateDone, tasks } = require('../taskUtil')
 const BlockFaces = require('prismarine-world').iterators.BlockFace
 const { Vec3 } = require('vec3')
