+++ lib/plugins/fishing.js
@@ -1,5 +1,5 @@
 const { Vec3 } = require('vec3')
-const { taskCreateDone, tasks } = require('../tasks')
+const { taskCreateDone, tasks } = require('../taskUtil')
 
 module.exports = inject
