+++ lib/plugins/inventory.js
@@ -2,7 +2,7 @@ const assert = require('assert')
 const { Vec3 } = require('vec3')
 const { once } = require('events')
 const { sleep, withTimeout } = require('../promise_utils')
-const { taskCreateDone, tasks } = require('../tasks')
+const { taskCreateDone, tasks } = require('../taskUtil')
 
 module.exports = inject
