+++ lib/plugins/physics.js
@@ -3,7 +3,7 @@ const assert = require('assert')
 const math = require('../math')
 const conv = require('../conversions')
 const { performance } = require('perf_hooks')
-const { taskCreateDone, tasks } = require('../tasks')
+const { taskCreateDone, tasks } = require('../taskUtil')
 
 const { Physics, PlayerState } = require('prismarine-physics')
