+++ lib/plugins/tasks.js
@@ -1,7 +1,6 @@
 const path = require('path')
-const { Task, ErrorTaskAlreadyRunning, ErrorTaskConflict } = require(path.join(__dirname, '../tasks.js'))
+const { Task, ErrorTaskAlreadyRunning, ErrorTaskConflict } = require(path.join(__dirname, '../taskUtil'))
 const TaskData = require(path.join(__dirname, '../task_data.json'))
-const { on } = require('events')
 const EventEmitter = require('events')
 const { onceWithCleanup } = require('../promise_utils')
 
@@ -61,7 +60,7 @@ function inject (bot) {
   async function taskWaitCompatibility (taskIdentifier) {
     const { status } = taskCompatibleWithRunningTask(taskIdentifier)
     if (status === true) return
-    onceWithCleanup(bot.tasks.taskUpdate, 'update', { 
+    onceWithCleanup(bot.tasks.taskUpdate, 'update', {
       checkCondition: () => {
         const { status } = taskCompatibleWithRunningTask(taskIdentifier)
         return status !== false
