+++ lib/promise_utils.js
@@ -1,4 +1,4 @@
-const { taskCreateBlank } = require('./tasks')
+const { taskCreateBlank } = require('./taskUtil')
 
 function sleep (ms) {
   return new Promise(resolve => setTimeout(resolve, ms))
