+++ lib/taskUtil.js
